从数据库读取字段

+++ src/main/java/com/shnu/work/repository/SystemInformationRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * @author Shinomiya Kaguya
  */
 @Repository
-public interface SystemInformationRepository extends CrudRepository<SystemInformationEntity, Integer> {
+public interface SystemInformationRepository extends CrudRepository<SystemInformationEntity, Long> {
 }
